backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2

# LLM & RAG
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
openai==1.47.0

# Vector Store
faiss-cpu==1.8.0

# Document Processing
PyMuPDF==1.24.10
Pillow==10.4.0
pytesseract==0.3.10

# ML / Classification
transformers==4.44.2
torch==2.4.1
sentencepiece==0.2.0

# Utilities
python-dateutil==2.9.0
aiofiles==24.1.0
python-dotenv==1.0.1
